# OtpAuthServices/bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "OtpAuthServices, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/BookTechnician/CreateBookTechnician": {
      "post": {
        "tags": [
          "BookTechnician"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/BookTechnician"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/BookTechnician/{id}": {
      "delete": {
        "tags": [
          "BookTechnician"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "BookTechnician"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/BookTechnician"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/BookTechnician/GetBookTechnician/{id}": {
      "get": {
        "tags": [
          "BookTechnician"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/BookTechnician/GetBookTechnicianDetailsForUserList": {
      "get": {
        "tags": [
          "BookTechnician"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "userId",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/BookTechnician/bookTechnicianEdit": {
      "post": {
        "tags": [
          "BookTechnician"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/PaymentRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/BookTechnician/GetBookTechnicianNotifications": {
      "get": {
        "tags": [
          "BookTechnician"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "category",
            "type": "string"
          },
          {
            "in": "query",
            "name": "pincode",
            "type": "string"
          },
          {
            "in": "query",
            "name": "technicianName",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Technician/{id}": {
      "put": {
        "tags": [
          "Technician"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Technician"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Technician/{technicianId}": {
      "get": {
        "tags": [
          "Technician"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "technicianId",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "BookTechnician": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "bookTechnicianId": {
          "type": "string"
        },
        "date": {
          "format": "date-time",
          "type": "string"
        },
        "customerName": {
          "type": "string"
        },
        "customerEmail": {
          "type": "string"
        },
        "address": {
          "type": "string"
        },
        "category": {
          "type": "string"
        },
        "status": {
          "type": "string"
        },
        "assignedTo": {
          "type": "string"
        },
        "technicianFullName": {
          "type": "string"
        },
        "customerId": {
          "type": "string"
        },
        "state": {
          "type": "string"
        },
        "statusCode": {
          "format": "int32",
          "type": "integer"
        },
        "district": {
          "type": "string"
        },
        "zipCode": {
          "type": "string"
        },
        "phoneNumber": {
          "type": "string"
        },
        "remarks": {
          "type": "string"
        },
        "rate": {
          "type": "string"
        },
        "discount": {
          "type": "string"
        },
        "moreInfo": {
          "type": "string"
        },
        "afterDiscount": {
          "type": "string"
        },
        "jobDescription": {
          "type": "string"
        },
        "paymentMode": {
          "type": "string"
        },
        "approvedAmount": {
          "type": "string"
        },
        "utrTransactionNumber": {
          "type": "string"
        },
        "technicianConfirmationCode": {
          "type": "string"
        },
        "orderId": {
          "type": "string"
        },
        "orderDate": {
          "type": "string"
        },
        "paidAmount": {
          "type": "string"
        },
        "transactionStatus": {
          "type": "string"
        },
        "transactionType": {
          "type": "string"
        },
        "invoiceId": {
          "type": "string"
        },
        "invoiceURL": {
          "type": "string"
        },
        "technicianPincode": {
          "type": "string"
        },
        "technicianName": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
...
    "PaymentRequest": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "orederId": {
          "type": "string"
        },
        "orderDate": {
          "type": "string"
        },
        "paidAmount": {
          "type": "string"
        },
        "transactionStatus": {
          "type": "string"
        },
        "transactionType": {
          "type": "string"
        },
        "utrNumber": {
          "type": "string"
        },
        "invoiceId": {
          "type": "string"
        },
        "invoiceURL": {
          "type": "string"
        }
      }
    },
...
    "Technician": {
      "type": "object",
      "properties": {
        "technicianId": {
          "format": "uuid",
          "type": "string"
        },
        "id": {
          "type": "string"
        },
        "technicianFullName": {
          "type": "string"
        },
        "panNumber": {
          "type": "string"
        },
        "numberOfTechnicians": {
          "format": "int32",
          "type": "integer"
        },
        "aadharNumber": {
          "type": "string"
        },
        "address": {
          "type": "string"
        },
        "state": {
          "type": "string"
        },
        "stateId": {
          "type": "string"
        },
        "district": {
          "type": "string"
        },
        "districtId": {
          "type": "string"
        },
        "zipCode": {
          "type": "string"
        },
        "phoneNumber": {
          "type": "string"
        },
        "alternativePhoneNumber": {
          "type": "string"
        },
        "phoneVerificationCode": {
          "type": "string"
        },
        "emailAddress": {
          "type": "string"
        },
        "emailVerificationCode": {
          "type": "string"
        },
        "category": {
          "type": "string"
        },
        "technicianPhotoId": {
          "type": "string"
        },
        "userId": {
          "format": "uuid",
          "type": "string"
        },
        "isApproved": {
          "type": "boolean"
        },
        "isRejected": {
          "type": "boolean"
        },
        "isPending": {
          "type": "boolean"
        },
        "status": {
          "type": "string"
        },
        "isActive": {
          "type": "boolean"
        }
      }
    },
...
  }
}